+++ udata_front/theme/gouvfr/datagouv-components/src/components/ReuseCard/ReuseCard.stories.ts
@@ -323,3 +323,25 @@ export const RouterReuseWithOrganization: StoryObj<typeof meta> = {
   decorators: [vueRouter(datasetRoutes)],
   args,
 };
+
+export const ReuseCardWithLongOrganizationName: StoryObj<typeof meta> = {
+  render: (args) => ({
+    components: { ReuseCard },
+    setup() {
+      return { args };
+    },
+    template: `<div class="fr-grid-row fr-grid-row--middle">
+                <div class="fr-col-6">
+                  <ReuseCard class="fr-mx-1v" v-bind="args"/>
+                </div>
+                <div class="fr-col-6">
+                  <ReuseCard class="fr-mx-1v" v-bind="args"/>
+                </div>
+              </div>`,
+  }),
+  decorators: [vueRouter(datasetRoutes)],
+  args: {
+    ...args,
+    reuse: {...args.reuse, owner: { ...args.reuse.owner, first_name: "Really really long first name", last_name: "Really really long last name" } },
+  },
+};